fix: Correct output in the export modal that is copy/pastable
Before this PR, the export modal output with hex values is missing quotes. So once you copy/paste the output to your code, you need to apply quotes around all the hex values.

And, the output with names are kebab-cased, so you need to transform them all to camelcase after copy/pasting code.

This PR produces correct output that is directly copy/pastable to the code.